fix(trusty-code): apply review findings — retry-scoped cap, attempts off-by-one, docs lint
Addresses code-review feedback on #2856 (fixes #2852):

- HIGH: the run-wide ceiling now counts FAILED engineer invocations
  (MAX_FAILED_INVOCATIONS) instead of every invocation. A run with several
  successful, distinct delegations never trips it — only invocations that
  actually failed feed the terminal ceiling. Total attempts are still
  tracked via `attempts()` for observability, but that counter governs
  nothing.
- MEDIUM: fixed the attempts()/report off-by-one — the terminal report now
  quotes `failed_invocations()` (the ceiling's real input) instead of a
  pre-incremented dispatch counter, so an operator is never told "13" when
  12 invocations actually failed. Updated the test that had pinned the
  wrong (off-by-one) value.
- LOW: corrected three stale `Test:` doc-comment pointers left over from
  the MAX_ENGINEER_INVOCATIONS -> MAX_FAILED_INVOCATIONS /
  run_wide_invocation_ceiling_latches_cap_reached -> failure_ceiling_latches_cap_reached
  renames (redelegation.rs:171,298; agent_loop/tests.rs:1731), and split
  redelegation.rs's inline test module into redelegation/tests.rs to keep
  the production file under the 500-SLOC cap.
- Repairs the "Test: doc-comment pointer lint" CI check
  (scripts/check_test_pointers.sh), which was failing on those three
  dangling pointers.
