+++ crates/trusty-code/src/run_task/tests.rs
@@ -940,6 +940,91 @@ fn assemble_report_maps_turn_cap_exceeded_with_deliverable_to_partial() {
     );
 }
 
+/// `assemble_report` maps a retry-exhausted run WITH a deliverable to
+/// `Partial`, labelled "re-delegation limit reached" (#2852).
+///
+/// Why: Post-#2852 this is arguably the most common real-world shape: retry
+/// exhaustion is now RECOVERABLE, so the PM keeps its turns and the loop ends
+/// on some other error while `retry_budget_exhausted` stays latched. Reporting
+/// keys off that flag rather than only `cap_reached` precisely so such runs
+/// keep their honest "the engineer was retried to exhaustion" diagnosis instead
+/// of silently degrading to an opaque `run_failure` — the exact regression
+/// #2852 set out to prevent. Nothing pinned this label's text, so it could have
+/// rotted freely; this test is that pin.
+/// What: Calls `assemble_report` directly with a signal that has ONLY
+/// `retry_budget_exhausted` latched (never `cap_reached`), a non-`TurnCap` PM
+/// error, and a before/after pair that differ. Assert `Partial`, and that the
+/// label names the re-delegation limit and the attempt count — NOT the
+/// invocation ceiling, which did not fire.
+/// Test: this test.
+#[test]
+fn assemble_report_maps_retry_exhausted_with_deliverable_to_partial() {
+    let params = RunTaskParams {
+        agent: "pm".into(),
+        task: "write hello.py".into(),
+        project: PathBuf::from("/tmp/does-not-matter"),
+        agents_dir: PathBuf::from("/tmp/does-not-matter-agents"),
+        engineer_model: None,
+        deadline_secs: None,
+    };
+    let transcript: super::SharedTranscript = Arc::new(Mutex::new(Vec::new()));
+
+    let before = super::diff::Snapshot::Files(std::collections::BTreeMap::new());
+    let mut after_map = std::collections::BTreeMap::new();
+    after_map.insert("hello.py".to_string(), "print('hi')".to_string());
+    let after = super::diff::Snapshot::Files(after_map);
+
+    // Deliberately NOT a turn cap: the point is that the retry-exhausted flag
+    // alone carries the label, independent of which loop error surfaced.
+    let pm_result: Result<AgentOutput, AgentLoopError> =
+        Err(AgentLoopError::Llm(LlmError::ApiError {
+            status: 500,
+            body: "upstream exploded".to_string(),
+        }));
+
+    let signal = RedelegationCapSignal::retry_exhausted_for_test();
+    let completion = EngineerCompletionSignal::new();
+    let report = super::assemble_report(
+        &params,
+        &transcript,
+        "openai/gpt-4o-mini",
+        "openai/gpt-4o-mini",
+        super::RunOutcome {
+            before,
+            after,
+            pm_result,
+        },
+        &signal,
+        &completion,
+    );
+
+    assert_eq!(
+        report.exit,
+        ExitCode::Partial,
+        "a retry-exhausted run that still produced a deliverable must map to Partial, \
+         not RunFailure (#2852), got label: {}",
+        report.task
+    );
+    assert!(
+        report.task.contains("re-delegation limit reached"),
+        "the label must name the re-delegation limit so the diagnosis is not lost, \
+         got: {}",
+        report.task
+    );
+    assert!(
+        report
+            .task
+            .contains(&super::redelegation::MAX_REDELEGATIONS.to_string()),
+        "the label must quote the per-delegation attempt count, got: {}",
+        report.task
+    );
+    assert!(
+        !report.task.contains("failure ceiling"),
+        "the run-wide ceiling never fired — the label must not claim it did, got: {}",
+        report.task
+    );
+}
+
 /// The mirror negative case: a `TurnCapExceeded` PM error with an EMPTY diff
 /// (no deliverable at all) stays `RunFailure` — the gate is purely "was work
 /// produced", per fix #4's requirement to not weaken the genuine-failure path.
@@ -1228,7 +1313,7 @@ impl LlmClientTrait for FirstEngineerCallWritesThenAllFail {
     }
 }
 
-/// Once the run-wide engineer-invocation ceiling latches, the PM must stop
+/// Once the run-wide engineer-failure ceiling latches, the PM must stop
 /// issuing `delegate_to_agent` calls (bounded well under its `max_turns` of 8)
 /// and the run must still end `Partial` with the deliverable preserved.
 ///
@@ -1238,19 +1323,28 @@ impl LlmClientTrait for FirstEngineerCallWritesThenAllFail {
 /// attempts … turn cap of 8 exceeded", 3 productive attempts + 7 wasted PM
 /// turns). #2852 narrowed WHICH condition may fire that unrecoverable hook to
 /// the one a fresh delegation cannot clear — the run-wide
-/// `MAX_ENGINEER_INVOCATIONS` ceiling — so this test now drives the ceiling
+/// `MAX_FAILED_INVOCATIONS` ceiling — so this test now drives the ceiling
 /// rather than a single delegation's retry exhaustion.
 ///
 /// This test previously asserted `pm_turns == 1`: that a PM whose FIRST
-/// delegation exhausted its retries never got a second turn. That assertion
-/// encoded the #2852 bug itself — it is exactly what killed L4 run-6, whose
-/// 4th delegation was the real build. Being allowed a further turn there is
-/// the fix, not a regression, so the scenario is rebuilt around the ceiling
-/// and the bound is now "well under max_turns" rather than "exactly 1".
+/// delegation exhausted its retries never got a second turn. That assertion is
+/// dropped as a deliberate POLICY change, and the record is worth stating
+/// precisely: the old scenario (one delegation whose engineer failed its own
+/// retries three times) was the cap's original, still-valid purpose — a
+/// genuinely struggling engineer — NOT run-6's bug, which was several separate
+/// SUCCESSFUL delegations starving a later one. So the old assertion was not
+/// itself defective, and an earlier version of this comment overstated the case
+/// by claiming it "encoded the bug". What actually changed is the policy: under
+/// #2852 that scenario now latches `retry_budget_exhausted` (recoverable), so
+/// the PM does get a second turn — which is the intended behaviour, because a
+/// fresh delegation with a full budget may well succeed. The scenario is
+/// therefore rebuilt around the ceiling, and the bound is now "well under
+/// max_turns" rather than "exactly 1".
 /// What: The engineer writes the deliverable on its first turn, then every
-/// engineer call fails retryably forever while the PM keeps delegating. Each
-/// delegation burns `MAX_REDELEGATIONS` invocations, so the ceiling latches on
-/// delegation 5. Assert: `exit == Partial`, the diff still names `hello.py`,
+/// engineer call fails retryably forever while the PM keeps delegating. That
+/// first success costs no ceiling budget (#2852 — only failures count);
+/// each subsequent delegation burns `MAX_REDELEGATIONS` failures, so the
+/// ceiling latches on delegation 5. Assert: `exit == Partial`, the diff still names `hello.py`,
 /// the report names the ceiling, and the PM stopped strictly before its own
 /// 8-turn cap — proof the stop signal, not the turn cap, ended the run.
 /// Test: this test.
@@ -1288,7 +1382,7 @@ async fn run_wide_ceiling_stops_the_pm_loop_and_ends_partial_promptly() {
         report.diff
     );
     assert!(
-        report.task.contains("engineer invocation ceiling reached"),
+        report.task.contains("engineer failure ceiling reached"),
         "the report must name the run-wide ceiling (#2852) as the terminal \
          condition, got: {}",
         report.task
@@ -1301,7 +1395,7 @@ async fn run_wide_ceiling_stops_the_pm_loop_and_ends_partial_promptly() {
          run, got {pm_turns} PM turns"
     );
     // Precise invocation bounding is asserted at the unit level in
-    // `redelegation::tests::run_wide_invocation_ceiling_latches_cap_reached`;
+    // `redelegation::tests::failure_ceiling_latches_cap_reached`;
     // here it is enough that the PM stopped delegating rather than spinning
     // out its whole turn budget on calls the ceiling would only refuse.
     assert!(
